Arrays for storing book list.
I created 3 arrays in UserProfile.java to store a list of books for the
user, along with 2 methods to access and test the arrays. The methods
are being called from AddBookServlet.java in the “do post” method

The methods and arrays are currently set to static because I am unsure
which class we want to create an instance of Userprofile.

I am however facing an issue where the values are not being inserted
into the array and the method to print the array in the console is not
consistently working. Could someone please take a look at it and see
what I am doing wrong.